+++ tests/test_gmf_output_fragment_runtime.py
@@ -41,6 +41,7 @@ struct EspAfe {
  Ring *fetch_output_ring_=&storage;
  std::atomic<uint32_t> fetch_timeout_{0}, output_ring_drop_{0},fetch_ok_{0},fetch_queue_frames_{0},fetch_queue_peak_{0};
  std::atomic<TaskHandle_t> pipeline_flush_waiter_{nullptr};
+ void apply_pending_gmf_vad_state_() {}
  void update_fetch_ring_free_pct_() {}
  esp_gmf_err_io_t gmf_output_release_(esp_gmf_payload_t *,int);
 };